requirements.txt
altair==5.1.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
catboost==1.2.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.5
contourpy==1.1.1
cycler==0.12.0
filelock==3.12.4
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.37
graphviz==0.20.1
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
scikit-image==0.21.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.27.1
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ultralytics==8.0.190
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
